requirements.txt
# FastAPI & Core
fastapi>=0.103.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic-settings>=2.0.3

# Database
SQLAlchemy>=2.0.20
psycopg2-binary>=2.9.7
neo4j>=5.12.0

# Security & Auth
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# AI & Graph Core
google-genai>=0.3.0
networkx>=3.1

# Vision Models (YOLO & Table Transformer)
ultralytics>=8.0.0 # YOLOv8
transformers>=4.33.0
torch>=2.0.1
torchvision>=0.15.2
Pillow>=10.0.0

# Audio & Video Models (Scene Detect & Whisper)
scenedetect[opencv]>=0.6.2
faster-whisper>=0.9.0